+++ src/plugins/jobacct_gather/linux/jobacct_gather_linux.c
@@ -118,7 +118,7 @@ static pthread_mutex_t reading_mutex = PTHREAD_MUTEX_INITIALIZER;
 static bool jobacct_shutdown = 0;
 static bool jobacct_suspended = 0;
 static List task_list = NULL;
-static uint32_t cont_id = (uint32_t)NO_VAL;
+static uint64_t cont_id = (uint64_t)NO_VAL;
 static bool pgid_plugin = false;
 
 /* Finally, pre-define all local routines. */
@@ -211,7 +211,7 @@ static void _get_process_data(void)
 	static int processing = 0;
 	long		hertz;
 
-	if(!pgid_plugin && cont_id == (uint32_t)NO_VAL) {
+	if (!pgid_plugin && (cont_id == (uint64_t)NO_VAL)) {
 		debug("cont_id hasn't been set yet not running poll");
 		return;
 	}
@@ -233,7 +233,7 @@ static void _get_process_data(void)
 		/* get only the processes in the proctrack container */
 		slurm_container_get_pids(cont_id, &pids, &npids);
 		if(!npids) {
-			debug4("no pids in this container %d", cont_id);
+			debug4("no pids in this container %lu", cont_id);
 			goto finished;
 		}
 		for (i = 0; i < npids; i++) {
@@ -766,18 +766,18 @@ extern void jobacct_gather_p_resume_poll()
 	jobacct_suspended = false;
 }
 
-extern int jobacct_gather_p_set_proctrack_container_id(uint32_t id)
+extern int jobacct_gather_p_set_proctrack_container_id(uint64_t id)
 {
 	if(pgid_plugin)
 		return SLURM_SUCCESS;
 
-	if(cont_id != (uint32_t)NO_VAL)
+	if (cont_id != (uint64_t)NO_VAL)
 		info("Warning: jobacct: set_proctrack_container_id: "
-		     "cont_id is already set to %d you are setting it to %d",
+		     "cont_id is already set to %lu you are setting it to %lu",
 		     cont_id, id);
-	if(id <= 0) {
+	if (id <= 0) {
 		error("jobacct: set_proctrack_container_id: "
-		      "I was given most likely an unset cont_id %d",
+		      "I was given most likely an unset cont_id %lu",
 		      id);
 		return SLURM_ERROR;
 	}
